Reasonably large sound overhaul to allow for using sound classes to generate tables of sounds without needing a backend, and more... [bugfix] logging names were adding in the sense of string contatenation rather than addition [refactor] base sound classes were split into their own module, and a bit of tidying was done to better handle their dependency checks and etc. Instead of BASE_CLASS being defined within the sounds module itself, you now use get_sound_class() which should read from prefs and do the thing properly [refactor] a dummy Sound class was made that will eventually be the parent class for all sounds (and then backend specific stuff will be added as a mixin. Currently the Sound class is not in the inheritance tree of the jack backend, but it will [lateral enhancement] added some methods in sound generation to accomodate the dummy class [logging] a few logging flags in the jack server [enhancement] Requires module! Module for specifying object requirements. just a stub for now, but will be where we go with v0.5.0 [enhancement] types module for declaring common and basic types. will eventually be used for handling units. remains to be seen how we will make data carrying structures. [partial bugfix] Added workarounds for prefs when multiprocessing fails, to use a dict instead of failing. Still needs some polish because of async exceptions, but better than before which would prevent import
