+++ docs/autopilot_docs.bib
@@ -35,7 +35,7 @@
 }
 
 @article{patonisFusionMethodCombining2018a,
-  title = {A {{Fusion Method}} for {{Combining Low}}-{{Cost IMU}}/{{Magnetometer Outputs}} for {{Use}} in {{Applications}} on {{Mobile Devices}}},
+  title = {A {{Fusion Method}} for {{Combining Low-Cost IMU}}/{{Magnetometer Outputs}} for {{Use}} in {{Applications}} on {{Mobile Devices}}},
   author = {Patonis, Photis and Patias, Petros and Tziavos, Ilias N. and Rossikopoulos, Dimitrios and Margaritis, Konstantinos G.},
   year = {2018},
   month = aug,
@@ -51,7 +51,7 @@
 }
 
 @article{slaneyEfficientImplementationPattersonHoldsworth1997,
-  title = {An {{Efficient Implementation}} of the {{Patterson}}-{{Holdsworth Auditory Filter Bank}}},
+  title = {An {{Efficient Implementation}} of the {{Patterson-Holdsworth Auditory Filter Bank}}},
   author = {Slaney, M.},
   year = {1997},
   journal = {undefined},
