+++ requirements.txt
@@ -11,3 +11,4 @@ scipy>=1.6.0
 pandas>=0.19.2
 tables>=3.4.2
 cffi
+validators
