+++ requirements/requirements_pilot.txt
@@ -17,3 +17,4 @@ tables
 cffi
 pigpio @ https://github.com/sneakers-the-rat/pigpio/tarball/master
 scipy>=1.6.0
+validators
